# backend/swagger.json
{
	"components": {
...
		"schemas": {
...
			"GetUsersResponse": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/UserDto"
						},
						"type": "array"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": true
			},
			"GetUsersByIdResponse": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/UserDto"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": true
			},
			"UpsertUserResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": true
			},
			"UpsertUserRequest": {
				"properties": {
					"id": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"active",
							"inactive"
						]
					},
					"roleId": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"fullName",
					"email"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"NewStatus": {
				"enum": [
					"draft",
					"publish"
				],
				"type": "string"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"name": "authorization",
				"type": "apiKey",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "nguoivietplus-api",
		"version": "0.0.0",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/admin-console/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetUsersResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": [
							"MANAGE_USERS_VIEW"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "UpsertUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UpsertUserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": [
							"MANAGE_USERS_ENABLE"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpsertUserRequest"
							}
						}
					}
				}
			}
		},
		"/admin-console/users/{userId}": {
			"get": {
				"operationId": "GetUsersById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetUsersByIdResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": [
							"MANAGE_USERS_DISABLE"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin-console/team": {
			"get": {
				"operationId": "GetTeam",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetUsersResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": [
							"MANAGE_MEMBER_VIEW"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "UpsertTeam",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UpsertUserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": [
							"MANAGE_MEMBER_ADD"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpsertUserRequest"
							}
						}
					}
				}
			}
		},
		"/admin-console/team/{userId}": {
			"get": {
				"operationId": "GetTeamById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetUsersByIdResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": [
							"MANAGE_MEMBER_VIEW"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteTeam",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": [
							"MANAGE_MEMBER_DELETE"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin-console/select/roles": {
			"get": {
				"operationId": "GetSelectRoles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SelectOptionResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": []
					}
				],
				"parameters": []
			}
		},
		"/admin-console/select/categories/{type}": {
			"get": {
				"operationId": "GetSelectCategories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SelectOptionResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "type",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/admin-console/categories": {
			"get": {
				"operationId": "GetCategories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetCategoriesResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "slug",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "UpsertCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UpsertCategoryResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": [
							"MANAGE_CATEGORY_ADD"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpsertCategoryRequest"
							}
						}
					}
				}
			}
		},
		"/admin-console/categories/{categoryId}": {
			"get": {
				"operationId": "GetCategoriesById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetCategoryByIdResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteCategory",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": [
							"MANAGE_CATEGORY_DELETE"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin-console/news": {
			"get": {
				"operationId": "GetNews",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PagedList_PagedNewsItemDto_"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"default": 30,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/NewStatus"
						}
					}
				]
			},
			"post": {
				"operationId": "UpsertNews",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NewsResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": [
							"MANAGE_POST_ADD"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NewsRequest"
							}
						}
					}
				}
			}
		},
		"/admin-console/news/{slug}": {
			"get": {
				"operationId": "GetNewsBySlug",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NewsDto"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "slug",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin-console/news/{newsId}": {
			"delete": {
				"operationId": "DeleteNews",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": [
							"MANAGE_POST_DELETE"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "newsId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin-console/news/{newsId}/change-status": {
			"put": {
				"operationId": "ChangeStatusNews",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": [
							"MANAGE_POST_EDIT"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "newsId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin-console/rao-vat": {
			"get": {
				"operationId": "GetRaoVats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetRaoVatResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "UpsertRaoVat",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RaoVatResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": [
							"RAOVAT_ADD"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RaoVatRequest"
							}
						}
					}
				}
			}
		},
		"/admin-console/rao-vat/{raoVatId}": {
			"get": {
				"operationId": "GetRaoVatsById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RaoVatResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "raoVatId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteRaoVat",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": [
							"RAOVAT_DELETE"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "raoVatId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/admin-console/menus": {
			"get": {
				"operationId": "GetMenus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetMenuResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "UpsertMenu",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": [
							"CONFIG_MENU_ADD"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MenuRequest"
							}
						}
					}
				}
			}
		},
		"/admin-console/menus/{menuId}": {
			"get": {
				"operationId": "GetMenusById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "menuId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteMenu",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": [
							"CONFIG_MENU_DELETE"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "menuId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin-console/configs/{key}": {
			"get": {
				"operationId": "GetConfigByKey",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConfigResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin-console/configs": {
			"post": {
				"operationId": "UpsertConfig",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConfigResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": [
							"CONFIG_SEO_EDIT"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ConfigRequest"
							}
						}
					}
				}
			}
		},
		"/admin-console/website": {
			"get": {
				"operationId": "GetWebsite",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WebsiteResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "UpsertWebsite",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WebsiteResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": [
							"CONFIG_SEO_EDIT"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/WebsiteRequest"
							}
						}
					}
				}
			}
		},
		"/admin-console/histories": {
			"get": {
				"operationId": "GetHistories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/HistoryDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": [
							"HISTORY_ACTIVITY_TEAM"
						]
					}
				],
				"parameters": []
			}
		},
...
		"/admin-console/reports/news-details-by/{userId}/months/{month}": {
			"get": {
				"operationId": "GetNewsDetailsOfMembers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetNewsDetailsOfMembersResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "month",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/admin-console/remove-images": {
			"delete": {
				"operationId": "RemoveImages",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "images",
						"required": true,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					}
				]
			}
		},
...
		"/news": {
			"get": {
				"operationId": "GetNews",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PagedList_PagedNewsItemDto_"
								}
							}
						}
					}
				},
				"tags": [
					"website"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"default": 30,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/NewStatus"
						}
					},
					{
						"in": "query",
						"name": "style",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"news1",
								"news2",
								"news3"
							]
						}
					},
					{
						"in": "query",
						"name": "isHotNews",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "menuSlug",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/rao-vat": {
			"get": {
				"operationId": "GetRaoVats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PagedList_PagedRaoVatItemDto_"
								}
							}
						}
					}
				},
				"tags": [
					"website"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"default": 30,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/NewStatus"
						}
					},
					{
						"in": "query",
						"name": "categorySlug",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000/api"
		}
	]
}